+++ backend/app/chat/chat_service.py
@@ -1,6 +1,5 @@
-from typing import Iterator
+from typing import AsyncIterator
 from pydantic import BaseModel
-from uuid import uuid4
 
 from google.api_core import exceptions
 from google.cloud import firestore
@@ -11,9 +10,10 @@ from ampf.base import AmpfBaseFactory, logger
 from app.agent.agent_model import Agent
 from gcp.gcp_file_storage import FileStorage
 from app.knowledge_base import KnowledgeBaseStorage
-from ai_model import AIModelFactory
+from haintech.ai import AiFactory
 
 from app.config import ServerConfig
+from haintech.ai.base.base_ai_text_embedding_model import BaseAITextEmbeddingModel
 from .chat_model import ChatSessionHeader, ChatSession
 from .message import ChatMessage, ChatMessageFile
 
@@ -32,19 +32,25 @@ class ChatService:
     """Service for chat."""
 
     def __init__(
-        self, factory: AmpfBaseFactory, file_storage: FileStorage, config: ServerConfig
+        self,
+        factory: AmpfBaseFactory,
+        ai_factory: AiFactory,
+        embedding_model: BaseAITextEmbeddingModel,
+        file_storage: FileStorage,
+        config: ServerConfig,
+        user_email: str,
     ):
-        self.ai_factory = AIModelFactory()
+        self.ai_factory = ai_factory
         self.role = ""
         self.storage = factory.create_storage(
             "ChatSessions", ChatSession, key_name="chat_session_id"
         )
         self.knowledge_base_storage = KnowledgeBaseStorage(
-            self.ai_factory,
-            embedding_model=config.knowledge_base.embedding_model,
+            embedding_model,
             embedding_search_limit=config.knowledge_base.embedding_search_limit,
         )
         self.file_storage = file_storage
+        self.user_email = user_email
         self._log = logger.get_logger(__name__)
 
     def get_answer(
@@ -59,14 +65,14 @@ class ChatService:
         out_history = [ChatMessage.from_content(m, {}) for m in chat.get_history()]
         return (ret, out_history)
 
-    def get_answer_async(
+    async def get_answer_async(
         self,
         message: ChatMessage,
         files: list[ChatMessageFile],
         agent: Agent = None,
         model_name: str = None,
         chat_session: ChatSession = None,
-    ) -> Iterator[StreamedEvent]:
+    ) -> AsyncIterator[StreamedEvent]:
         """Get an answer from the model."""
         file_names = {}
         in_history = []
@@ -80,38 +86,45 @@ class ChatService:
         if agent:
             model_name = agent.model_name
         try:
-            context = self.get_context(message.content, agent)
+            context = await self.get_context(message.content, agent)
             ai_agent = AIAgent(model_name=model_name, system_instruction=context)
             chat = ai_agent.start_chat(history=in_history)
             parts = [message.content]
             # Iterate over the user (session) files
             for file in files:
                 # Move the file to the chat session directory
-                chat_blob_name = f"chat-{chat_session.chat_session_id}/{str(uuid4())}"
-                blob = self.file_storage.move_blob(file.name, chat_blob_name)
-                # Create a part with the file content
-                blob_dict = BlobDict(
-                    mime_type=file.mime_type, data=blob.download_as_string()
+                # file_name = f"users/{self.user_email}/session_files/{file.name}"
+                chat_blob_name = f"users/{self.user_email}/chats/{chat_session.chat_session_id}/files/{file.name}"
+                # file.url = chat_blob_name
+                self._log.debug("Moving file %s to %s", file.name, chat_blob_name)
+                self.file_storage.move_blob(
+                    f"users/{self.user_email}/session_files/{file.name}", chat_blob_name
                 )
+                blob_data = self.file_storage.download_blob(chat_blob_name)
+                # Create a part with the file content
+                blob_dict = BlobDict(mime_type=file.mime_type, data=blob_data)
                 parts.append(blob_dict)
             content = ContentDict(role="user", parts=parts)
+            chat_session.history.append(
+                ChatMessage(author="user", content=message.content, files=files)
+            )
+            self._log.debug("Sending message: %s", content)
             responses = chat.send_message_streaming(content)
             for response in responses:
+                self._log.debug("Received response: %s", response)
                 if response.text:
                     yield StreamedEvent(type="text", value=response.text)
                 # await asyncio.sleep(0.1)
-            out_history = [
-                ChatMessage.from_content(m, file_names) for m in chat.get_history()
-            ]
-            chat_session.history = out_history
+            out_message = ChatMessage.from_content(chat.get_history()[-1], file_names)
+            chat_session.history.append(out_message)
             if not chat_session.summary:
-                chat_session.summary = out_history[0].content
+                chat_session.summary = chat_session.history[0].content
             self.storage.save(chat_session)
         except Exception as e:
             self._log.exception("Error in get_answer_async: %s", e)
             yield StreamedEvent(type=f"error:{type(e).__name__}", value=str(e))
 
-    def get_context(self, text: str, agent: Agent = None) -> str:
+    async def get_context(self, text: str, agent: Agent = None) -> str:
         """Get the context of the chat session."""
         if agent:
             context = agent.system_prompt + "\n\n"
@@ -122,7 +135,7 @@ class ChatService:
         else:
             context = ""
             keywords = None
-        neartest = self.knowledge_base_storage.find_nearest(f"{text}", keywords)
+        neartest = await self.knowledge_base_storage.find_nearest(f"{text}", keywords)
         for n in neartest:
             context += "\n\n# " + n.title + "\n" + n.content + "\n\n"
         return context
@@ -162,13 +175,16 @@ class ChatService:
     async def delete_chat(self, chat_session_id: str, user: str) -> None:
         """Delete chat history by id."""
         chat_session = self.storage.get(chat_session_id)
+        if not chat_session:
+            self._log.warning("Chat session not found: %s", chat_session_id)
+            return
         if chat_session.user != user:
             raise ChatSessionUserError()
         chat_session = self.storage.get(chat_session_id)
         for message in chat_session.history:
             for file in message.files:
                 try:
-                    self.file_storage.delete("/".join(file.url.split("/")[3:]))
+                    self.file_storage.delete(file.name)
                 except exceptions.NotFound:
                     pass
         return self.storage.delete(chat_session_id)