+++ backend/app/dependencies.py
@@ -5,6 +5,7 @@ from dotenv import load_dotenv
 from fastapi import Depends
 from fastapi.security import OAuth2PasswordBearer
 
+from haintech.ai import AiFactory
 from ampf.auth import TokenPayload, AuthService, InsufficientPermissionsError
 from ampf.base import AmpfBaseFactory, BaseEmailSender, SmtpEmailSender, EmailTemplate
 from ampf.gcp import AmpfGcpFactory
@@ -15,19 +16,20 @@ from app.user.user_service import UserService
 from app.config import ServerConfig
 from app.agent import AgentService
 from app.chat import ChatService
+from haintech.ai.base.base_ai_text_embedding_model import BaseAITextEmbeddingModel
 
 
 load_dotenv()
 
 
-def get_server_config() -> ServerConfig:
+async def get_server_config() -> ServerConfig:
     return ServerConfig()
 
 
 ServerConfigDep = Annotated[ServerConfig, Depends(get_server_config)]
 
 
-def get_factory() -> AmpfBaseFactory:
+async def get_factory() -> AmpfBaseFactory:
     return AmpfGcpFactory()
 
 
@@ -37,21 +39,21 @@ oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/login")
 AuthTokenDep = Annotated[str, Depends(oauth2_scheme)]
 
 
-def user_service_dep(factory: FactoryDep) -> UserService:
+async def user_service_dep(factory: FactoryDep) -> UserService:
     return UserService(factory)
 
 
 UserServceDep = Annotated[UserService, Depends(user_service_dep)]
 
 
-def get_email_sender(conf: ServerConfigDep) -> BaseEmailSender:
+async def get_email_sender(conf: ServerConfigDep) -> BaseEmailSender:
     return SmtpEmailSender(**dict(conf.smtp))
 
 
 EmailSenderServiceDep = Annotated[BaseEmailSender, Depends(get_email_sender)]
 
 
-def auth_service_dep(
+async def auth_service_dep(
     factory: FactoryDep,
     email_sender_service: EmailSenderServiceDep,
     conf: ServerConfigDep,
@@ -72,14 +74,14 @@ def auth_service_dep(
 AuthServiceDep = Annotated[AuthService, Depends(auth_service_dep)]
 
 
-def decode_token(auth_service: AuthServiceDep, token: AuthTokenDep):
+async def decode_token(auth_service: AuthServiceDep, token: AuthTokenDep):
     return auth_service.decode_token(token)
 
 
 TokenPayloadDep = Annotated[TokenPayload, Depends(decode_token)]
 
 
-def get_user_email(token_payload: TokenPayloadDep) -> str:
+async def get_user_email(token_payload: TokenPayloadDep) -> str:
     """Returns the current user's ID from the session."""
     return token_payload.email
 
@@ -100,14 +102,16 @@ class Authorize:
             raise InsufficientPermissionsError()
 
 
-def get_agent_service(config: ServerConfigDep, factory: FactoryDep) -> AgentService:
+async def get_agent_service(
+    config: ServerConfigDep, factory: FactoryDep
+) -> AgentService:
     return AgentService(config, factory)
 
 
 AgentServiceDep = Annotated[AgentService, Depends(get_agent_service)]
 
 
-def get_file_service(
+async def get_file_service(
     config: ServerConfigDep, factory: FactoryDep, user_email: UserEmailDep
 ) -> FileService:
     return FileService(config, factory, user_email)
@@ -116,10 +120,40 @@ def get_file_service(
 FileServiceDep = Annotated[FileService, Depends(get_file_service)]
 
 
-def get_chat_service(
-    factory: FactoryDep, server_config: ServerConfigDep, file_service: FileServiceDep
+async def get_ai_factory() -> AiFactory:
+    return AiFactory()
+
+
+AiFactoryDep = Annotated[AiFactory, Depends(get_ai_factory)]
+
+
+async def get_ai_text_embedding_model(
+    ai_factory: AiFactoryDep, config: ServerConfigDep
+):
+    return ai_factory.get_text_embedding_model(config.knowledge_base.embedding_model)
+
+
+EmbeddingModelDep = Annotated[
+    BaseAITextEmbeddingModel, Depends(get_ai_text_embedding_model)
+]
+
+
+async def get_chat_service(
+    factory: FactoryDep,
+    ai_factory: AiFactoryDep,
+    embedding_model: EmbeddingModelDep,
+    server_config: ServerConfigDep,
+    file_service: FileServiceDep,
+    user_email: UserEmailDep,
 ) -> ChatService:
-    return ChatService(factory, file_service.storage, server_config)
+    return ChatService(
+        factory,
+        ai_factory,
+        embedding_model,
+        file_service.storage,
+        server_config,
+        user_email,
+    )
 
 
 ChatServiceDep = Annotated[ChatService, Depends(get_chat_service)]