+++ backend/app/main.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI, Request
 from fastapi.staticfiles import StaticFiles
 
 from app.logging_conf import setup_logging
-from app.routers import auth, chats, config, files
+from app.routers import auth, chats, config, files, upgrade
 
 from app.dependencies import ServerConfigDep
 from .routers import users, agents, knowledge_base
@@ -22,7 +22,7 @@ app.include_router(prefix="/api/files", router=files.router)
 app.include_router(prefix="/api/chats", router=chats.router)
 app.include_router(prefix="/api/agents", router=agents.router)
 app.include_router(prefix="/api/knowledge-base", router=knowledge_base.router)
-
+app.include_router(prefix="/api/upgrade", router=upgrade.router)
 
 # @app.get("/api/auth")
 # async def auth_google(
@@ -47,12 +47,12 @@ async def user_get(user_service: users.UserServiceDep, request: Request):
 
 
 @app.get("/api/ping")
-def ping() -> None:
+async def ping() -> None:
     """Just for keep container alive"""
 
 
 @app.get("/api/models")
-def models_get_all(config: ServerConfigDep) -> list[str]:
+async def models_get_all(config: ServerConfigDep) -> list[str]:
     return [m.strip() for m in config.get("models").split(",")]