+++ backend/app/routers/chats_message.py
@@ -1,7 +1,7 @@
 import logging
-from typing import Iterator
 from fastapi import APIRouter
-from fastapi.responses import StreamingResponse
+
+from ampf import JsonStreamingResponse
 
 from app.chat.chat_model import ChatSession
 from app.dependencies import (
@@ -11,7 +11,6 @@ from app.dependencies import (
     UserEmailDep,
     FileServiceDep,
 )
-from app.chat.chat_service import StreamedEvent
 from app.chat.message.message_model import ChatMessage, ChatMessageFile
 
 
@@ -23,7 +22,7 @@ _log = logging.getLogger(__name__)
 
 
 @router.post("", responses={200: {"content": {"text/event-stream": {}}}})
-def post_message_async(
+async def post_message_async(
     message: ChatMessage,
     config: ServerConfigDep,
     user_email: UserEmailDep,
@@ -47,19 +46,11 @@ def post_message_async(
         agent_obj = agent_service.create_default(
             user_email, model_name=model if model else config.default_model
         )
-    responses = chat_service.get_answer_async(
-        agent=agent_obj,
-        chat_session=chat_session,
-        message=message,
-        files=files,
-    )
-
-    def handle_history(responses: Iterator[StreamedEvent]):
-        for i, r in enumerate(responses):
-            comma = "," if i > 0 else ""
-            yield f"{comma}{r.model_dump_json()}\n"
-
-    return StreamingResponse(
-        handle_history(responses),
-        media_type="text/event-stream",
+    return JsonStreamingResponse(
+        chat_service.get_answer_async(
+            agent=agent_obj,
+            chat_session=chat_session,
+            message=message,
+            files=files,
+        ),
     )