+++ backend/tests/app/features/test_user_service.py
@@ -5,7 +5,11 @@ from app.user import UserService, User
 
 @pytest.fixture
 def user_service(factory):
-    return UserService(factory)
+    service = UserService(factory)
+    yield service
+    # Clean up
+    service.storage_new.drop()
+    service.storage_old.drop()
 
 
 def test_put_get_and_delete_user(user_service):
@@ -20,3 +24,27 @@ def test_put_get_and_delete_user(user_service):
     user_service.delete(email)
     # Then: The user cannot be retrieved from the storage
     assert user_service.get(email) is None
+
+
+def test_upgrade_not_exists(user_service):
+    # Given: A user in the old storage
+    email = "jasio@wp.pl"
+    user = User(email=email, password="test", name="Jasio")
+    user_service.storage_old.put(email, user)
+    # When: The upgrade is run
+    user_service.upgrade()
+    # Then: The user is in the new storage
+    assert user_service.storage_new.get(email) == user
+
+
+def test_upgrade_exists(user_service):
+    # Given: A user in both storages
+    email = "jasio@wp.pl"
+    user_o = User(email=email, password="test", name="Jasio")
+    user_service.storage_old.put(email, user_o)
+    user_n = User(email=email, password="test", name="Jan")
+    user_service.storage_new.put(email, user_n)
+    # When: The upgrade is run
+    user_service.upgrade()
+    # Then: The user in the new storage is not changed
+    assert user_service.storage_new.get(email) == user_n