+++ backend/tests/conftest.py
@@ -1,10 +1,11 @@
-from typing import Optional
+from typing import List, Optional
 import pytest
 
 from ampf.auth.auth_model import TokenExp
 from ampf.base import BaseEmailSender
 from ampf.storage_in_memory import AmpfInMemoryFactory
 from app.config import DefaultUserConfig, ServerConfig
+from haintech.ai.base.base_ai_text_embedding_model import BaseAITextEmbeddingModel
 
 
 @pytest.fixture
@@ -45,10 +46,15 @@ def email_sender():
 
 
 @pytest.fixture
-def test_config():
+def user_email() -> str:
+    return "test@test.com"
+
+
+@pytest.fixture
+def test_config(user_email):
     """Return a test configuration."""
     return ServerConfig(
-        default_user=DefaultUserConfig(email="test@test.com", password="test"),
+        default_user=DefaultUserConfig(email=user_email, password="test"),
         file_storage_bucket="vertex-ai-chat-dev-unit-tests",
     )
 
@@ -74,3 +80,13 @@ def tokens(factory, client):
 @pytest.fixture
 def auth_header(tokens):
     yield {"Authorization": f"Bearer {tokens['access_token']}"}
+
+
+class MockAITextEmbeddingModel(BaseAITextEmbeddingModel):
+    async def get_embedding(self, text: str) -> List[float]:
+        return [0.1] * 256
+
+
+@pytest.fixture
+def embedding_model():
+    return MockAITextEmbeddingModel()