+++ backend/tests/e2e/conftest.py
@@ -1,4 +1,5 @@
 import json
+import logging
 from typing import Dict
 from fastapi import Response
 from fastapi.testclient import TestClient
@@ -6,7 +7,12 @@ import pytest
 
 
 from ampf.auth.auth_model import TokenExp
-from app.dependencies import get_email_sender, get_factory, get_server_config
+from app.dependencies import (
+    get_ai_text_embedding_model,
+    get_email_sender,
+    get_factory,
+    get_server_config,
+)
 from app.main import app
 
 
@@ -51,13 +57,17 @@ class AuthClient(TestClient):
 
 
 @pytest.fixture
-def client(factory, email_sender, test_config):
+def client(factory, embedding_model, email_sender, test_config):
     """Create a FastAPI test client where the app is the main FastAPI app."""
+
+    logging.getLogger("app").setLevel(logging.DEBUG)
+    logging.getLogger("ampf").setLevel(logging.DEBUG)
+    logging.getLogger("gcp").setLevel(logging.DEBUG)
 
     app.dependency_overrides[get_factory] = lambda: factory
     app.dependency_overrides[get_email_sender] = lambda: email_sender
     app.dependency_overrides[get_server_config] = lambda: test_config
-
+    app.dependency_overrides[get_ai_text_embedding_model] = lambda: embedding_model
     client = AuthClient(app)
     # Clear token_black_list
     factory.create_compact_storage("token_black_list", TokenExp, "token").drop()